decodeToken function
